feat(dapp): Refactor auctions related codebase to 'auctions module' folder structure
