+++ dapp/src/auctions/hooks/useGetAuctionMetadata.ts
@@ -4,8 +4,8 @@
 import { useQuery } from '@tanstack/react-query';
 
 import { useIotaNamesClient } from '@/contexts';
-import { createAuctionMetadataQuery } from '@/lib/auction/metadata';
 
+import { createAuctionMetadataQuery } from '../lib/metadata';
 import { useAuctionHouse } from './useAuctionHouse';
 
 export function useGetAuctionMetadata(domainName: string) {
