+++ dapp/src/auctions/lib/bcs.ts
