+++ dapp/src/auctions/lib/metadata.ts
