+++ dapp/src/auctions/lib/types.ts
