+++ dapp/src/auctions/lib/utils.ts
