+++ dapp/src/auctions/services/IotaNamesIndexerClient.ts
