+++ dapp/src/components/dialogs/index.ts
@@ -1,5 +1,4 @@
 // Copyright (c) 2025 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './AuctionBidDialog';
 export * from './PurchaseNameDialog';
