+++ dapp/src/hooks/index.ts
@@ -8,5 +8,4 @@ export * from './useRegisterNameTransaction';
 export * from './useRegistrationNfts';
 export * from './useTheme';
 export * from './usePriceList';
-export * from './auction/useGetAddressAuctionHistory';
 export * from './queryKey';
